Add etcd{,_test}.go
