+++ config/config.go
@@ -3,11 +3,6 @@
 // file, which is used by config-bootstrapper and cloud-config-server.
 package config
 
-import (
-	"fmt"
-	"strings"
-)
-
 // Cluster configures a cluster, which includes: (1) a
 // bootstrapper machine, (2) the Kubernetes cluster.
 type Cluster struct {
@@ -39,40 +34,4 @@ type Node struct {
 	CephMonitor bool   `yaml:"ceph_monitor"`
 	KubeMaster  bool   `yaml:"kube_master"`
 	EtcdMember  bool   `yaml:"etcd_member"`
-}
-
-// Join is defined as a method of Cluster, so can be called in
-// templates.
-func (c Cluster) Join(s []string) string {
-	return strings.Join(s, ", ")
-}
-
-// Hostname is defined as a method of Node, so can be call in
-// template.
-func (n Node) Hostname() string {
-	return strings.ToUpper(strings.Replace(n.MAC, ":", "-", -1))
-}
-
-// Mac is defined as a method of Node, so can be called in template.
-func (n Node) Mac() string {
-	return strings.ToUpper(n.MAC)
-}
-
-// InitialEtcdCluster derives the value of command line parameter
-// --initial_cluster of etcd from Cluter.Nodes and Node.EtcdMember.
-// NOTE: Every node in the cluster will have a etcd daemon running --
-// either as a member or as a proxy.
-func (c Cluster) InitialEtcdCluster() string {
-	var ret []string
-	for _, n := range c.Nodes {
-		if n.EtcdMember {
-			name := n.Hostname()
-			addr := n.Hostname()
-			if len(n.IP) > 0 {
-				addr = n.IP // No need for DNS then.
-			}
-			ret = append(ret, fmt.Sprintf("%s=http://%s:2380", name, addr))
-		}
-	}
-	return strings.Join(ret, ",")
 }
